backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.2
psycopg2-binary==2.9.11
